Update assignment1.html
The activity that was completed was the activity 5 which was tracing internet traffic. Within the activity five I was asked to review the how-to geek article about how to utilize traceroute in order to identify problems along the network it. In order to find out more about this traceroute command I had to use the tracert on my windows operating system to trace the route from me to en.wikiversity .org. I also utilized the website traceroute online to understand more about this activity.  I learned that I could use traceroute to identify problems in ta network. I also learned the behind-the-scenes f how traceroute operates with routers in order to exploit time to live servers. I envision being able to utilize this in my future career as a cybersecurity officer.